+++ android/app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
          additional functionality it is fine to subclass or reimplement
          FlutterApplication and put your custom class here. -->
 
-    <application android:usesCleartextTraffic="true" android:name="io.flutter.app.FlutterApplication" android:label="ZY_Player_flutter" android:icon="@mipmap/ic_launcher">
+    <application android:requestLegacyExternalStorage="true"  android:usesCleartextTraffic="true" android:name="io.flutter.app.FlutterApplication" android:label="ZY_Player_flutter" android:icon="@mipmap/ic_launcher">
         <activity android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection|fontScale|screenLayout|density|uiMode" android:hardwareAccelerated="true" android:launchMode="singleTop" android:screenOrientation="portrait" android:name=".MainActivity" android:theme="@style/LaunchTheme" android:windowSoftInputMode="adjustResize">
             <!-- Specify that the launch screen should continue being displayed -->
             <!-- until Flutter renders its first frame. -->
