+++ android/app/src/main/res/drawable-night/launch_background.xml
@@ -5,10 +5,11 @@
         <color android:color="#FF18191A"/>
     </item>
 
+
     <!-- You can insert your own image assets here -->
-    <!-- <item>
-        <bitmap
-            android:gravity="center"
-            android:src="@mipmap/launch_image" />
-    </item> -->
+<!--    <item>-->
+<!--        <bitmap-->
+<!--            android:gravity="center"-->
+<!--            android:src="@mipmap/ic_launcher" />-->
+<!--    </item>-->
 </layer-list>
