+++ lib/login/page/update_password_page.dart
@@ -1,13 +1,13 @@
-import 'package:flutter/material.dart';
-import 'package:ZY_Player_flutter/util/change_notifier_manage.dart';
+import 'package:ZY_Player_flutter/login/widgets/my_text_field.dart';
 import 'package:ZY_Player_flutter/res/resources.dart';
 import 'package:ZY_Player_flutter/routes/fluro_navigator.dart';
+import 'package:ZY_Player_flutter/util/change_notifier_manage.dart';
 import 'package:ZY_Player_flutter/util/toast.dart';
 import 'package:ZY_Player_flutter/util/utils.dart';
-import 'package:ZY_Player_flutter/widgets/app_bar.dart';
+import 'package:ZY_Player_flutter/widgets/my_app_bar.dart';
 import 'package:ZY_Player_flutter/widgets/my_button.dart';
 import 'package:ZY_Player_flutter/widgets/my_scroll_view.dart';
-import 'package:ZY_Player_flutter/login/widgets/my_text_field.dart';
+import 'package:flutter/material.dart';
 
 /// design/1注册登录/index.html#artboard13
 class UpdatePasswordPage extends StatefulWidget {
