+++ lib/provider/theme_provider.dart
@@ -1,10 +1,10 @@
 import 'dart:ui';
 
+import 'package:ZY_Player_flutter/common/common.dart';
+import 'package:ZY_Player_flutter/res/resources.dart';
 import 'package:flustars/flustars.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:ZY_Player_flutter/common/common.dart';
-import 'package:ZY_Player_flutter/res/resources.dart';
 
 class ThemeProvider extends ChangeNotifier {
   static const Map<ThemeMode, String> themes = {ThemeMode.dark: 'Dark', ThemeMode.light: 'Light', ThemeMode.system: 'System'};
@@ -12,8 +12,12 @@ class ThemeProvider extends ChangeNotifier {
   bool _loadingState = false;
   bool get loadingState => _loadingState;
 
-  void setloadingState(bool state) {
+  String _loadingText = "";
+  String get loadingText => _loadingText;
+
+  void setloadingState(bool state, [String text]) {
     _loadingState = state;
+    _loadingText = text;
     notifyListeners();
   }
