+++ lib/setting/page/account_manager_page.dart
@@ -1,9 +1,9 @@
-import 'package:flutter/material.dart';
 import 'package:ZY_Player_flutter/login/login_router.dart';
 import 'package:ZY_Player_flutter/routes/fluro_navigator.dart';
-import 'package:ZY_Player_flutter/widgets/app_bar.dart';
 import 'package:ZY_Player_flutter/widgets/click_item.dart';
 import 'package:ZY_Player_flutter/widgets/load_image.dart';
+import 'package:ZY_Player_flutter/widgets/my_app_bar.dart';
+import 'package:flutter/material.dart';
 
 /// design/8设置/index.html#artboard1
 class AccountManagerPage extends StatefulWidget {
