+++ lib/setting/page/setting_page.dart
@@ -1,11 +1,11 @@
-import 'package:flustars/flustars.dart';
-import 'package:flutter/material.dart';
 import 'package:ZY_Player_flutter/common/common.dart';
 import 'package:ZY_Player_flutter/res/resources.dart';
 import 'package:ZY_Player_flutter/routes/fluro_navigator.dart';
 import 'package:ZY_Player_flutter/setting/widgets/exit_dialog.dart';
-import 'package:ZY_Player_flutter/widgets/app_bar.dart';
 import 'package:ZY_Player_flutter/widgets/click_item.dart';
+import 'package:ZY_Player_flutter/widgets/my_app_bar.dart';
+import 'package:flustars/flustars.dart';
+import 'package:flutter/material.dart';
 
 import '../setting_router.dart';
