+++ lib/util/utils.dart
@@ -1,7 +1,7 @@
-import 'package:common_utils/common_utils.dart';
-import 'package:flutter/material.dart';
 import 'package:ZY_Player_flutter/util/theme_utils.dart';
 import 'package:ZY_Player_flutter/util/toast.dart';
+import 'package:common_utils/common_utils.dart';
+import 'package:flutter/material.dart';
 import 'package:keyboard_actions/keyboard_actions.dart' as keyboard;
 import 'package:keyboard_actions/keyboard_actions_config.dart';
 import 'package:url_launcher/url_launcher.dart';
@@ -65,41 +65,6 @@ class Utils {
               )),
     );
   }
-}
-
-/// 默认dialog背景色为半透明黑色，这里修改源码改为透明
-Future<T> showTransparentDialog<T>({
-  @required BuildContext context,
-  bool barrierDismissible = true,
-  WidgetBuilder builder,
-}) {
-  final ThemeData theme = Theme.of(context, shadowThemeOnly: true);
-  return showGeneralDialog(
-    context: context,
-    pageBuilder: (BuildContext buildContext, Animation<double> animation, Animation<double> secondaryAnimation) {
-      final Widget pageChild = Builder(builder: builder);
-      return SafeArea(
-        child: Builder(builder: (BuildContext context) {
-          return theme != null ? Theme(data: theme, child: pageChild) : pageChild;
-        }),
-      );
-    },
-    barrierDismissible: barrierDismissible,
-    barrierLabel: MaterialLocalizations.of(context).modalBarrierDismissLabel,
-    barrierColor: const Color(0x00FFFFFF),
-    transitionDuration: const Duration(milliseconds: 150),
-    transitionBuilder: _buildMaterialDialogTransitions,
-  );
-}
-
-Widget _buildMaterialDialogTransitions(BuildContext context, Animation<double> animation, Animation<double> secondaryAnimation, Widget child) {
-  return FadeTransition(
-    opacity: CurvedAnimation(
-      parent: animation,
-      curve: Curves.easeOut,
-    ),
-    child: child,
-  );
 }
 
 Future<T> showElasticDialog<T>({
