+++ pubspec.yaml
@@ -34,8 +34,6 @@ dependencies:
   dio: ^3.0.10
   # https://github.com/ReactiveX/rxdart
   rxdart: ^0.24.1
-  # Dart 常用工具类库 https://github.com/Sky24n/common_utils
-  common_utils: 1.2.1
   # Flutter 常用工具类库 https://github.com/Sky24n/flustars
   flustars: ^0.3.3
   # Flutter 轮播图 https://github.com/best-flutter/flutter_swiper
@@ -47,13 +45,13 @@ dependencies:
   # 侧滑删除 https://github.com/letsar/flutter_slidable
   flutter_slidable: ^0.5.7
   # WebView插件 https://github.com/flutter/plugins/tree/master/packages/webview_flutter
-  webview_flutter: ^1.0.5
+  webview_flutter: ^1.0.7
   # 处理键盘事件 https://github.com/diegoveloper/flutter_keyboard_actions
   keyboard_actions: ^3.3.1+1
   # 路由框架 https://github.com/theyakka/fluro
   fluro: ^1.7.7
   # 图片缓存 https://github.com/renefloor/flutter_cached_network_image
-  cached_network_image: ^2.3.3
+  cached_network_image: ^2.4.1
   # 格式化String https://github.com/Naddiseo/dart-sprintf
   sprintf: ^5.0.0
   # 状态管理 https://github.com/rrousselGit/provider
@@ -70,16 +68,20 @@ dependencies:
   flutter_update_dialog: ^1.0.0
   #动画
   flutter_staggered_animations: "^0.1.2"
-  ##
-  device_info: ^1.0.0
-  battery: ^1.0.7
   # 视频推送
   flutter_dlna: ^0.0.1+1
   flutter_picker:
     git: git://github.com/yangyxd/flutter_picker.git
+  #dialog
+  giffy_dialog: 1.8.0
+  #权限
+  permission_handler: 5.0.1+1
 
   # 统计
   janalytics_fluttify: 0.3.2+1
+  #二维码
+  qr_flutter: 3.2.0
+  image_pickers: 1.0.8+7
 
 dev_dependencies:
   build_runner: ^1.10.4
